+++ src/controllers/comments/index.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
 import httpStatus from 'http-status';
 import log4js from 'log4js';
-import type { CommentSaveDto } from 'src/dto/comment/commentSaveDto.type';
+import type { CommentSaveDto } from 'src/comments/types/commentSaveDto.type';
 import {
     countCommentsByArticleId,
     listCommentsByArticleId,
