--- src/dto/student/studentDetailsDto.ts
@@ -1,12 +1,0 @@
-import { Address } from 'src/dto/address/address';
-
-export interface StudentDetailsDto {
-	name: string;
-	surname: string;
-	groupId: string;
-	birthDate: Date;
-	phoneNumbers: string[];
-	address?: Address;
-	createdAt: Date;
-	updatedAt: Date;
-}
