--- src/dto/student/studentInfoDto.ts
@@ -1,5 +1,0 @@
-export interface StudentInfoDto {
-	_id: string;
-	fullName: string;
-	groupId: string;
-}
