+++ src/test/controllers/comment.test.ts
@@ -3,8 +3,8 @@ import chai from 'chai';
 import chaiHttp from 'chai-http';
 import express from 'express';
 import sinon from 'sinon';
-import type { CommentSaveDto } from 'src/dto/comment/commentSaveDto.type';
-import Comment from 'src/model/comment';
+import Comment from 'src/comments/comment.model';
+import type { CommentSaveDto } from 'src/comments/types/commentSaveDto.type';
 import routers from 'src/routers/comments';
 import * as commentService from 'src/services/comments';
 import { comments, seed } from '../config/commentsConfig';
